refactor(view): view 를 입력과 출력 기준으로 RequestView 및 ResultView 로 재구성
