+++ src/main/kotlin/view/ResultView.kt
@@ -1,5 +1,6 @@
 package view
 
+import domain.constant.DEALER_STAND_CONDITION
 import domain.person.Dealer
 import domain.person.Person
 import domain.person.Player
@@ -7,18 +8,39 @@ import domain.result.GameResult
 import domain.result.OutCome
 
 object ResultView {
-    private const val INITIAL_CARDS_SCRIPT = "%s 카드: %s - 결과: %s"
     private const val FINAL_OUTCOME_SCRIPT = "## 최종 승패"
     private const val DEALER_SCRIPT = "딜러: "
-    fun printPersonsCards(dealer: Dealer, players: List<Player>) {
-        printPersonCards(dealer)
-        players.forEach { printPersonCards(it) }
+    private const val SHARE_TWO_CARDS_SCRIPT = "%s와 %s에게 2장의 카드를 나누었습니다."
+    private const val INITIAL_CARDS_SCRIPT = "%s 카드: %s"
+    private const val RESULT_CARDS_SCRIPT = "%s 카드: %s - 결과: %s"
+    private const val DEALER_ONE_MORE_CARD_SCRIPT = "딜러는 ${DEALER_STAND_CONDITION}이하라 한장의 카드를 더 받았습니다."
+    private const val DEALER_NO_MORE_CARD_SCRIPT = "딜러는 ${DEALER_STAND_CONDITION}초과라 한장의 카드를 받지 않습니다."
+
+    fun printInitialSetting(players: List<Player>, dealer: Dealer) {
+        println()
+        println(SHARE_TWO_CARDS_SCRIPT.format(dealer.name, players.joinToString(", ") { it.name }))
+        println(INITIAL_CARDS_SCRIPT.format(dealer.name, dealer.showOneCard().joinToString { it.toString() }))
+        players.forEach { printPlayerCards(it) }
         println()
     }
 
-    private fun printPersonCards(person: Person) {
+    fun printPlayerCards(person: Person) {
+        println(INITIAL_CARDS_SCRIPT.format(person.name, person.cards.joinToString(",") { it.toString() }))
+    }
+
+    fun printDealerGetMoreCard() = println(DEALER_ONE_MORE_CARD_SCRIPT + "\n")
+
+    fun printDealerNoMoreCard() = println(DEALER_NO_MORE_CARD_SCRIPT + "\n")
+
+    fun printPersonsCardsResult(dealer: Dealer, players: List<Player>) {
+        printPersonCardsResult(dealer)
+        players.forEach { printPersonCardsResult(it) }
+        println()
+    }
+
+    private fun printPersonCardsResult(person: Person) {
         println(
-            INITIAL_CARDS_SCRIPT.format(
+            RESULT_CARDS_SCRIPT.format(
                 person.name,
                 person.cards.joinToString(",") { it.toString() },
                 person.getTotalCardNumber(),
